DeleteLaboratoryCommand

+++ LIMS.Application/Handlers/Laboratory/LaboratoryCommandHandlers/DeleteLaboratoryCommandHandler.cs
@@ -21,14 +21,16 @@ namespace LIMS.Application.Handlers.Laboratory.LaboratoryCommandHandlers
         public async Task<Unit> Handle(DeleteLaboratoryCommand request, CancellationToken cancellationToken)
         {
             var entityToDelete = await _labQueryRepository.GetAsyncById(request.Id, DataTables.LaboratoryTable, DataColumns.LaboratoryId);
-
-            var laboratoryEntity = AutoMapperConfiguration.Mapper.Map<Domain.Models.Laboratory>(entityToDelete);
-            if (laboratoryEntity == null)
+            if (entityToDelete != null)
             {
-                throw new ApplicationException("Unable to map due to an issue with mapper.");
+                var laboratoryEntity = AutoMapperConfiguration.Mapper.Map<Domain.Models.Laboratory>(entityToDelete);
+                if (laboratoryEntity == null)
+                {
+                    throw new ApplicationException("Unable to map due to an issue with mapper.");
+                }
+                else
+                    await _labCommandRepository.DeleteAsync(laboratoryEntity);
             }
-
-            await _labCommandRepository.DeleteAsync(laboratoryEntity);
             return Unit.Value;
         }
     }
